Refactor the homepage
