+++ yarn-error.log
@@ -1,0 +1,110 @@
+Arguments: 
+  C:\Program Files\nodejs\node.exe C:\Program Files (x86)\Yarn\bin\yarn.js install
+
+PATH: 
+  C:\Users\LENOVO PC\bin;C:\Program Files\Git\mingw64\bin;C:\Program Files\Git\usr\local\bin;C:\Program Files\Git\usr\bin;C:\Program Files\Git\usr\bin;C:\Program Files\Git\mingw64\bin;C:\Program Files\Git\usr\bin;C:\Users\LENOVO PC\bin;C:\Python34;C:\Python34\Scripts;C:\ProgramData\Oracle\Java\javapath;C:\Program Files (x86)\Intel\iCLS Client;C:\Program Files\Intel\iCLS Client;C:\Windows\system32;C:\Windows;C:\Windows\System32\Wbem;C:\Windows\System32\WindowsPowerShell\v1.0;C:\Program Files\Intel\Intel(R) Management Engine Components\DAL;C:\Program Files\Intel\Intel(R) Management Engine Components\IPT;C:\Program Files (x86)\Intel\Intel(R) Management Engine Components\DAL;C:\Program Files (x86)\Intel\Intel(R) Management Engine Components\IPT;C:\Program Files (x86)\Intel\OpenCL SDK\2.0\bin\x86;C:\Program Files (x86)\Intel\OpenCL SDK\2.0\bin\x64;%QUARTUS_ROOTDIR%\bin;C:\Program Files\MATLAB\R2010b\runtime\win64;C:\Program Files\MATLAB\R2010b\bin;C:\Program Files\PuTTY;C:\Program Files (x86)\Windows Kits\8.1\Windows Performance Toolkit;C:\Program Files\MySQL\MySQL Utilities 1.6;C:\Program Files\nodejs;C:\Program Files (x86)\Yarn\bin;C:\Ruby24-x64\bin;C:\Ruby22-x64\bin;C:\Program Files (x86)\Microsoft VS Code\bin;C:\Program Files\Java\jdk1.8.0_121\bin;C:\Program Files\PostgreSQL\9.6\bin;C:\Program Files\PostgreSQL\9.6\lib;C:\Program Files\Heroku\bin;C:\Users\LENOVO PC\AppData\Roaming\npm;C:\Users\LENOVO PC\AppData\Local\Yarn\bin;C:\Users\LENOVO PC\AppData\Local\Programs\Microsoft VS Code\bin;C:\Program Files\Git\usr\bin\vendor_perl;C:\Program Files\Git\usr\bin\core_perl
+
+Yarn version: 
+  1.9.4
+
+Node version: 
+  8.11.1
+
+Platform: 
+  win32 x64
+
+Trace: 
+  Error: https://registry.yarnpkg.com/rxjs/-/rxjs-6.6.3.tgz: ESOCKETTIMEDOUT
+      at ClientRequest.<anonymous> (C:\Program Files (x86)\Yarn\lib\cli.js:138095:19)
+      at Object.onceWrapper (events.js:313:30)
+      at emitNone (events.js:106:13)
+      at ClientRequest.emit (events.js:208:7)
+      at TLSSocket.emitTimeout (_http_client.js:711:34)
+      at Object.onceWrapper (events.js:313:30)
+      at emitNone (events.js:106:13)
+      at TLSSocket.emit (events.js:208:7)
+      at TLSSocket.Socket._onTimeout (net.js:420:8)
+      at ontimeout (timers.js:482:11)
+
+npm manifest: 
+  {
+    "name": "newfrontend",
+    "version": "1.0.0",
+    "description": "A simple game app built with react",
+    "main": "index.js",
+    "scripts": {
+      "test": "jest --coverage",
+      "start": "babel-node server.js"
+    },
+    "keywords": [
+      "React"
+    ],
+    "dependencies": {
+      "@babel/polyfill": "7.2.5",
+      "axios": "0.18.0",
+      "dotenv-webpack": "1.7.0",
+      "enzyme": "3.8.0",
+      "enzyme-adapter-react-16": "1.7.1",
+      "express": "4.16.4",
+      "html-webpack-plugin": "3.2.0",
+      "prop-types": "15.6.2",
+      "react": "16.7.0",
+      "react-dom": "16.7.0",
+      "react-redux": "6.0.0",
+      "react-router-dom": "4.3.1",
+      "redux": "4.0.1",
+      "redux-thunk": "2.3.0",
+      "toastr": "2.1.2",
+      "webpack": "4.26.1",
+      "webpack-dev-middleware": "1.6.1",
+      "webpack-hot-middleware": "2.10.0"
+    },
+    "devDependencies": {
+      "@babel/cli": "7.2.3",
+      "@babel/core": "7.1.2",
+      "@babel/node": "7.2.2",
+      "@babel/plugin-proposal-class-properties": "7.3.0",
+      "@babel/plugin-proposal-object-rest-spread": "7.3.1",
+      "@babel/plugin-transform-runtime": "7.2.0",
+      "@babel/preset-env": "7.1.0",
+      "@babel/preset-react": "7.0.0",
+      "@babel/register": "7.0.0",
+      "babel-eslint": "10.0.1",
+      "babel-jest": "23.6.0",
+      "babel-loader": "8.0.4",
+      "compression": "^1.7.3",
+      "coveralls": "3.0.2",
+      "css-loader": "0.23.1",
+      "dot-env": "0.0.1",
+      "eslint": "5.12.0",
+      "eslint-config-airbnb": "17.1.0",
+      "eslint-loader": "2.1.1",
+      "eslint-plugin-import": "2.14.0",
+      "eslint-plugin-jsx": "0.0.2",
+      "eslint-plugin-jsx-a11y": "6.1.2",
+      "eslint-plugin-react": "7.12.3",
+      "eventsource-polyfill": "0.9.6",
+      "extract-text-webpack-plugin": "1.0.1",
+      "file-loader": "^3.0.1",
+      "jest": "24.0.0",
+      "jsdom": "13.2.0",
+      "mocha": "5.2.0",
+      "nock": "8.0.0",
+      "redux-immutable-state-invariant": "1.2.3",
+      "redux-mock-store": "1.5.3",
+      "sinon": "7.2.3",
+      "style-loader": "0.13.1",
+      "uglifyjs-webpack-plugin": "2.1.1",
+      "url-loader": "0.5.7",
+      "webpack-cli": "3.2.1",
+      "webpack-dev-server": "3.1.14"
+    },
+    "author": "Ademola Hussain",
+    "license": "MIT"
+  }
+
+yarn manifest: 
+  No manifest
+
+Lockfile: 
+  No lockfile
